Convert 'Arnold searches from the home page' to a feature spec
Running this in a real browser as opposed to rack-test exposed a bug
where the state param in the form action was being discarded by Chrome
resulting in all archived petitions being searched and not just the
published ones as expected. Also the order was different - I suspect
the cucumber matcher wasn't asserting the order of the results.